[EX] Adds optional dependencies
Compiling UXID in another project was raising the following:

warning: Ecto.ParameterizedType.__using__/1 defined in application :ecto is
used by the current application but the current application does not
directly depend on :ecto. To fix this, you must do one of:

  ...

  3. In case you don't want to add a requirement to :ecto, you may optionally
  skip this warning by adding [xref: [exclude: Ecto.ParameterizedType]] to your
  "def project" in mix.exs

This applies the suggestion by adding the xfref exclude.